refactor: switch Permission to direct self-reexport imports
