+++ packages/opencode/src/effect/app-runtime.ts
@@ -21,7 +21,7 @@ import { Agent } from "@/agent/agent"
 import { Skill } from "@/skill"
 import { Discovery } from "@/skill/discovery"
 import { Question } from "@/question"
-import { Permission } from "@/permission"
+import { Permission } from "@/permission/permission"
 import { Todo } from "@/session/todo"
 import { Session } from "@/session"
 import { SessionStatus } from "@/session/status"
