--- packages/opencode/src/permission/index.ts
@@ -1,1 +1,0 @@
-export * as Permission from "./permission"
