+++ packages/opencode/src/permission/permission.ts
@@ -321,3 +321,5 @@ export function disabled(tools: string[], ruleset: Ruleset): Set<string> {
 }
 
 export const defaultLayer = layer.pipe(Layer.provide(Bus.layer))
+
+export * as Permission from "./permission"
