+++ packages/opencode/src/server/instance/permission.ts
@@ -2,7 +2,7 @@ import { Hono } from "hono"
 import { describeRoute, validator, resolver } from "hono-openapi"
 import z from "zod"
 import { AppRuntime } from "@/effect/app-runtime"
-import { Permission } from "@/permission"
+import { Permission } from "@/permission/permission"
 import { PermissionID } from "@/permission/schema"
 import { errors } from "../error"
 import { lazy } from "../../util/lazy"
