+++ packages/opencode/src/server/instance/session.ts
@@ -19,7 +19,7 @@ import { Agent } from "../../agent/agent"
 import { Snapshot } from "@/snapshot"
 import { Command } from "../../command"
 import { Log } from "../../util"
-import { Permission } from "@/permission"
+import { Permission } from "@/permission/permission"
 import { PermissionID } from "@/permission/schema"
 import { ModelID, ProviderID } from "@/provider/schema"
 import { errors } from "../error"
