+++ packages/opencode/src/session/processor.ts
@@ -3,7 +3,7 @@ import * as Stream from "effect/Stream"
 import { Agent } from "@/agent/agent"
 import { Bus } from "@/bus"
 import { Config } from "@/config"
-import { Permission } from "@/permission"
+import { Permission } from "@/permission/permission"
 import { Plugin } from "@/plugin"
 import { Snapshot } from "@/snapshot"
 import * as Session from "./session"
