+++ packages/opencode/src/session/prompt.ts
@@ -35,7 +35,7 @@ import { SessionSummary } from "./summary"
 import { NamedError } from "@opencode-ai/shared/util/error"
 import { SessionProcessor } from "./processor"
 import { Tool } from "@/tool"
-import { Permission } from "@/permission"
+import { Permission } from "@/permission/permission"
 import { SessionStatus } from "./status"
 import { LLM } from "./llm"
 import { Shell } from "@/shell/shell"
