+++ packages/opencode/src/skill/skill.ts
@@ -9,7 +9,7 @@ import { Bus } from "@/bus"
 import { InstanceState } from "@/effect"
 import { Flag } from "@/flag/flag"
 import { Global } from "@/global"
-import { Permission } from "@/permission"
+import { Permission } from "@/permission/permission"
 import { AppFileSystem } from "@opencode-ai/shared/filesystem"
 import { Config } from "../config"
 import { ConfigMarkdown } from "../config"
