+++ packages/opencode/test/session/compaction.test.ts
@@ -11,7 +11,7 @@ import { SessionCompaction } from "../../src/session/compaction"
 import { Token } from "../../src/util"
 import { Instance } from "../../src/project/instance"
 import { Log } from "../../src/util"
-import { Permission } from "../../src/permission"
+import { Permission } from "../../src/permission/permission"
 import { Plugin } from "../../src/plugin"
 import { provideTmpdirInstance, tmpdir } from "../fixture/fixture"
 import { Session as SessionNs } from "../../src/session"
