+++ packages/opencode/test/session/processor-effect.test.ts
@@ -6,7 +6,7 @@ import type { Agent } from "../../src/agent/agent"
 import { Agent as AgentSvc } from "../../src/agent/agent"
 import { Bus } from "../../src/bus"
 import { Config } from "../../src/config"
-import { Permission } from "../../src/permission"
+import { Permission } from "../../src/permission/permission"
 import { Plugin } from "../../src/plugin"
 import { Provider } from "../../src/provider"
 import { ModelID, ProviderID } from "../../src/provider/schema"
