+++ packages/opencode/test/session/prompt-effect.test.ts
@@ -10,7 +10,7 @@ import { Config } from "../../src/config"
 import { FileTime } from "../../src/file/time"
 import { LSP } from "../../src/lsp"
 import { MCP } from "../../src/mcp"
-import { Permission } from "../../src/permission"
+import { Permission } from "../../src/permission/permission"
 import { Plugin } from "../../src/plugin"
 import { Provider as ProviderSvc } from "../../src/provider"
 import { Env } from "../../src/env"
