+++ packages/opencode/test/tool/skill.test.ts
@@ -3,7 +3,7 @@ import { Effect, Layer } from "effect"
 import { afterEach, describe, expect } from "bun:test"
 import path from "path"
 import { pathToFileURL } from "url"
-import type { Permission } from "../../src/permission"
+import type { Permission } from "../../src/permission/permission"
 import type { Tool } from "../../src/tool"
 import { Instance } from "../../src/project/instance"
 import { SkillTool } from "../../src/tool/skill"
